Extensions/WeatherData/WeatherData/requirements.txt
aiohttp==3.10.10
# debugpy

--find-links file:///C:/ProgramData/Beckhoff/PythonPackages
beckhoff-tchmi-extensionapi  # must be installed via TcPkg by the Python.Beckhoff.tchmi.extensionapi package.
